# [Java] Optimize putIntAscii/putLongAscii methods