--- a/agrona/src/main/java/org/agrona/AsciiEncoding.java
+++ b/agrona/src/main/java/org/agrona/AsciiEncoding.java
@@ -42,6 +42,23 @@ public final class AsciiEncoding
      */
     public static final byte ZERO = '0';
 
+    /**
+     * Lookup table used for encoding ints/longs as ASCII characters.
+     */
+    public static final byte[] ASCII_DIGITS = new String(new char[]
+    {
+        '0', '0', '0', '1', '0', '2', '0', '3', '0', '4', '0', '5', '0', '6', '0', '7', '0', '8', '0', '9',
+        '1', '0', '1', '1', '1', '2', '1', '3', '1', '4', '1', '5', '1', '6', '1', '7', '1', '8', '1', '9',
+        '2', '0', '2', '1', '2', '2', '2', '3', '2', '4', '2', '5', '2', '6', '2', '7', '2', '8', '2', '9',
+        '3', '0', '3', '1', '3', '2', '3', '3', '3', '4', '3', '5', '3', '6', '3', '7', '3', '8', '3', '9',
+        '4', '0', '4', '1', '4', '2', '4', '3', '4', '4', '4', '5', '4', '6', '4', '7', '4', '8', '4', '9',
+        '5', '0', '5', '1', '5', '2', '5', '3', '5', '4', '5', '5', '5', '6', '5', '7', '5', '8', '5', '9',
+        '6', '0', '6', '1', '6', '2', '6', '3', '6', '4', '6', '5', '6', '6', '6', '7', '6', '8', '6', '9',
+        '7', '0', '7', '1', '7', '2', '7', '3', '7', '4', '7', '5', '7', '6', '7', '7', '7', '8', '7', '9',
+        '8', '0', '8', '1', '8', '2', '8', '3', '8', '4', '8', '5', '8', '6', '8', '7', '8', '8', '8', '9',
+        '9', '0', '9', '1', '9', '2', '9', '3', '9', '4', '9', '5', '9', '6', '9', '7', '9', '8', '9', '9'
+    }).getBytes(US_ASCII);
+
     private static final byte[] MIN_INT_DIGITS = "2147483648".getBytes(US_ASCII);
     private static final byte[] MAX_INT_DIGITS = "2147483647".getBytes(US_ASCII);
 
@@ -141,9 +158,9 @@ public final class AsciiEncoding
      * @param cs     to parse.
      * @param index  at which the number begins.
      * @param length of the encoded number in characters.
-     * @throws AsciiNumberFormatException if <code>cs</code> is not an int value
-     * @throws IndexOutOfBoundsException if parsing results in access outside string boundaries, or length is negative
      * @return the parsed value.
+     * @throws AsciiNumberFormatException if <code>cs</code> is not an int value
+     * @throws IndexOutOfBoundsException  if parsing results in access outside string boundaries, or length is negative
      */
     public static int parseIntAscii(final CharSequence cs, final int index, final int length)
     {
@@ -181,9 +198,9 @@ public final class AsciiEncoding
      * @param cs     to parse.
      * @param index  at which the number begins.
      * @param length of the encoded number in characters.
-     * @throws AsciiNumberFormatException if <code>cs</code> is not a long value
-     * @throws IndexOutOfBoundsException if parsing results in access outside string boundaries, or length is negative
      * @return the parsed value.
+     * @throws AsciiNumberFormatException if <code>cs</code> is not a long value
+     * @throws IndexOutOfBoundsException  if parsing results in access outside string boundaries, or length is negative
      */
     public static long parseLongAscii(final CharSequence cs, final int index, final int length)
     {