--- a/agrona/src/main/java/org/agrona/AsciiEncoding.java
+++ b/agrona/src/main/java/org/agrona/AsciiEncoding.java
@@ -45,7 +45,7 @@ public final class AsciiEncoding
     /**
      * Lookup table used for encoding ints/longs as ASCII characters.
      */
-    public static final byte[] ASCII_DIGITS = new String(new char[]
+    public static final byte[] ASCII_DIGITS = new byte[]
     {
         '0', '0', '0', '1', '0', '2', '0', '3', '0', '4', '0', '5', '0', '6', '0', '7', '0', '8', '0', '9',
         '1', '0', '1', '1', '1', '2', '1', '3', '1', '4', '1', '5', '1', '6', '1', '7', '1', '8', '1', '9',
@@ -57,7 +57,7 @@ public final class AsciiEncoding
         '7', '0', '7', '1', '7', '2', '7', '3', '7', '4', '7', '5', '7', '6', '7', '7', '7', '8', '7', '9',
         '8', '0', '8', '1', '8', '2', '8', '3', '8', '4', '8', '5', '8', '6', '8', '7', '8', '8', '8', '9',
         '9', '0', '9', '1', '9', '2', '9', '3', '9', '4', '9', '5', '9', '6', '9', '7', '9', '8', '9', '9'
-    }).getBytes(US_ASCII);
+    };
 
     private static final byte[] MIN_INT_DIGITS = "2147483648".getBytes(US_ASCII);
     private static final byte[] MAX_INT_DIGITS = "2147483647".getBytes(US_ASCII);