--- a/agrona/src/main/java/org/agrona/ExpandableArrayBuffer.java
+++ b/agrona/src/main/java/org/agrona/ExpandableArrayBuffer.java
@@ -1174,12 +1174,25 @@ public class ExpandableArrayBuffer implements MutableDirectBuffer
 
         ensureCapacity(index, length);
 
-        while (i >= 0)
+        final byte[] dest = byteArray;
+        while (quotient >= 100)
         {
-            final int remainder = quotient % 10;
-            quotient = quotient / 10;
-            byteArray[i + start] = (byte)(ZERO + remainder);
-            i--;
+            final int position = (quotient % 100) << 1;
+            quotient /= 100;
+            dest[i + start] = ASCII_DIGITS[position + 1];
+            dest[i - 1 + start] = ASCII_DIGITS[position];
+            i -= 2;
+        }
+
+        if (quotient < 10)
+        {
+            dest[i + start] = (byte)(ZERO + quotient);
+        }
+        else
+        {
+            final int position = quotient << 1;
+            dest[i + start] = ASCII_DIGITS[position + 1];
+            dest[i - 1 + start] = ASCII_DIGITS[position];
         }
 
         return length;
@@ -1202,13 +1215,25 @@ public class ExpandableArrayBuffer implements MutableDirectBuffer
         ensureCapacity(index, length);
 
         int quotient = value;
-        while (i >= 0)
+        final byte[] dest = byteArray;
+        while (quotient >= 100)
         {
-            final int remainder = quotient % 10;
-            quotient = quotient / 10;
-            byteArray[i + index] = (byte)(ZERO + remainder);
+            final int position = (quotient % 100) << 1;
+            quotient /= 100;
+            dest[i + index] = ASCII_DIGITS[position + 1];
+            dest[i - 1 + index] = ASCII_DIGITS[position];
+            i -= 2;
+        }
 
-            i--;
+        if (quotient < 10)
+        {
+            dest[i + index] = (byte)(ZERO + quotient);
+        }
+        else
+        {
+            final int position = quotient << 1;
+            dest[i + index] = ASCII_DIGITS[position + 1];
+            dest[i - 1 + index] = ASCII_DIGITS[position];
         }
 
         return length;
@@ -1269,13 +1294,50 @@ public class ExpandableArrayBuffer implements MutableDirectBuffer
         ensureCapacity(index, length);
 
         long quotient = value;
-        while (i >= 0)
+        final byte[] dest = byteArray;
+        while (quotient >= 100000000)
         {
-            final long remainder = quotient % 10;
-            quotient = quotient / 10;
-            byteArray[i + index] = (byte)(ZERO + remainder);
+            final int lastEightDigits = (int)(quotient % 100000000);
+            quotient /= 100000000;
 
-            i--;
+            final int upperPart = lastEightDigits / 10000;
+            final int lowerPart = lastEightDigits % 10000;
+
+            final int u1 = (upperPart / 100) << 1;
+            final int u2 = (upperPart % 100) << 1;
+            final int l1 = (lowerPart / 100) << 1;
+            final int l2 = (lowerPart % 100) << 1;
+
+            i -= 8;
+
+            dest[index + i + 1] = ASCII_DIGITS[u1];
+            dest[index + i + 2] = ASCII_DIGITS[u1 + 1];
+            dest[index + i + 3] = ASCII_DIGITS[u2];
+            dest[index + i + 4] = ASCII_DIGITS[u2 + 1];
+            dest[index + i + 5] = ASCII_DIGITS[l1];
+            dest[index + i + 6] = ASCII_DIGITS[l1 + 1];
+            dest[index + i + 7] = ASCII_DIGITS[l2];
+            dest[index + i + 8] = ASCII_DIGITS[l2 + 1];
+        }
+
+        while (quotient >= 100)
+        {
+            final int position = (int)((quotient % 100) << 1);
+            quotient /= 100;
+            dest[index + i] = ASCII_DIGITS[position + 1];
+            dest[index + i - 1] = ASCII_DIGITS[position];
+            i -= 2;
+        }
+
+        if (quotient < 10)
+        {
+            dest[index + i] = (byte)(ZERO + quotient);
+        }
+        else
+        {
+            final int position = (int)(quotient << 1);
+            dest[index + i] = ASCII_DIGITS[position + 1];
+            dest[index + i - 1] = ASCII_DIGITS[position];
         }
 
         return length;
@@ -1314,12 +1376,50 @@ public class ExpandableArrayBuffer implements MutableDirectBuffer
 
         ensureCapacity(index, length);
 
-        while (i >= 0)
+        final byte[] dest = byteArray;
+        while (quotient >= 100000000)
         {
-            final long remainder = quotient % 10L;
-            quotient = quotient / 10L;
-            byteArray[i + start] = (byte)(ZERO + remainder);
-            i--;
+            final int lastEightDigits = (int)(quotient % 100000000);
+            quotient /= 100000000;
+
+            final int upperPart = lastEightDigits / 10000;
+            final int lowerPart = lastEightDigits % 10000;
+
+            final int u1 = (upperPart / 100) << 1;
+            final int u2 = (upperPart % 100) << 1;
+            final int l1 = (lowerPart / 100) << 1;
+            final int l2 = (lowerPart % 100) << 1;
+
+            i -= 8;
+
+            dest[start + i + 1] = ASCII_DIGITS[u1];
+            dest[start + i + 2] = ASCII_DIGITS[u1 + 1];
+            dest[start + i + 3] = ASCII_DIGITS[u2];
+            dest[start + i + 4] = ASCII_DIGITS[u2 + 1];
+            dest[start + i + 5] = ASCII_DIGITS[l1];
+            dest[start + i + 6] = ASCII_DIGITS[l1 + 1];
+            dest[start + i + 7] = ASCII_DIGITS[l2];
+            dest[start + i + 8] = ASCII_DIGITS[l2 + 1];
+        }
+
+        while (quotient >= 100)
+        {
+            final int position = (int)((quotient % 100) << 1);
+            quotient /= 100;
+            dest[start + i] = ASCII_DIGITS[position + 1];
+            dest[start + i - 1] = ASCII_DIGITS[position];
+            i -= 2;
+        }
+
+        if (quotient < 10)
+        {
+            dest[start + i] = (byte)(ZERO + quotient);
+        }
+        else
+        {
+            final int position = (int)(quotient << 1);
+            dest[start + i] = ASCII_DIGITS[position + 1];
+            dest[start + i - 1] = ASCII_DIGITS[position];
         }
 
         return length;