This PR uses lookup table approach to optimize `putIntAscii`, `putNaturalIntAscii`, `putLongAscii` and `putNaturalLongAscii` methods by dividing by `100` and looking up two digits at a time. I'm see up to **23%** performance improvement.

This work was inspired by https://github.com/miloyip/itoa-benchmark and https://www.slideshare.net/andreialexandrescu1/three-optimization-tips-for-c?next_slideshow=1.

Here is the results I got:
- Before:
```
Benchmark                                       (value)  Mode  Cnt   Score   Error  Units
UnsafeBufferPutIntAsciiBenchmark            -2147483648  avgt   30   6.543 ± 0.018  ns/op
UnsafeBufferPutIntAsciiBenchmark                      0  avgt   30   3.268 ± 0.015  ns/op
UnsafeBufferPutIntAsciiBenchmark                  -9182  avgt   30   7.612 ± 0.040  ns/op
UnsafeBufferPutIntAsciiBenchmark               27085146  avgt   30  11.929 ± 0.010  ns/op
UnsafeBufferPutIntAsciiBenchmark             -123456789  avgt   30  13.511 ± 0.043  ns/op
UnsafeBufferPutIntAsciiBenchmark             2147483647  avgt   30  14.069 ± 0.028  ns/op
UnsafeBufferPutLongAsciiBenchmark  -9223372036854775808  avgt   30   6.699 ± 0.027  ns/op
UnsafeBufferPutLongAsciiBenchmark                     0  avgt   30   2.490 ± 0.010  ns/op
UnsafeBufferPutLongAsciiBenchmark                 -9182  avgt   30   8.620 ± 0.013  ns/op
UnsafeBufferPutLongAsciiBenchmark              97385146  avgt   30  12.600 ± 0.042  ns/op
UnsafeBufferPutLongAsciiBenchmark     -6180362504315475  avgt   30  24.890 ± 0.065  ns/op
UnsafeBufferPutLongAsciiBenchmark   9223372036854775807  avgt   30  28.601 ± 0.129  ns/op
```
- After:
```
Benchmark                                       (value)  Mode  Cnt   Score   Error  Units
UnsafeBufferPutIntAsciiBenchmark            -2147483648  avgt   30   6.564 ± 0.035  ns/op
UnsafeBufferPutIntAsciiBenchmark                      0  avgt   30   3.162 ± 0.007  ns/op
UnsafeBufferPutIntAsciiBenchmark                  -9182  avgt   30   6.114 ± 0.013  ns/op
UnsafeBufferPutIntAsciiBenchmark               27085146  avgt   30   9.443 ± 0.035  ns/op
UnsafeBufferPutIntAsciiBenchmark             -123456789  avgt   30  11.573 ± 0.040  ns/op
UnsafeBufferPutIntAsciiBenchmark             2147483647  avgt   30  11.244 ± 0.024  ns/op
UnsafeBufferPutLongAsciiBenchmark  -9223372036854775808  avgt   30   8.089 ± 0.027  ns/op
UnsafeBufferPutLongAsciiBenchmark                     0  avgt   30   3.280 ± 0.013  ns/op
UnsafeBufferPutLongAsciiBenchmark                 -9182  avgt   30   8.206 ± 0.038  ns/op
UnsafeBufferPutLongAsciiBenchmark              97385146  avgt   30  11.116 ± 0.027  ns/op
UnsafeBufferPutLongAsciiBenchmark     -6180362504315475  avgt   30  19.083 ± 0.054  ns/op
UnsafeBufferPutLongAsciiBenchmark   9223372036854775807  avgt   30  21.832 ± 0.050  ns/op
```

Here are also the intermediate results:
<details><summary>Lookup table</summary>

```
Benchmark                                       (value)  Mode  Cnt   Score   Error  Units
UnsafeBufferPutIntAsciiBenchmark            -2147483648  avgt   30   6.555 ± 0.029  ns/op
UnsafeBufferPutIntAsciiBenchmark                      0  avgt   30   3.257 ± 0.024  ns/op
UnsafeBufferPutIntAsciiBenchmark                  -9182  avgt   30   7.220 ± 0.012  ns/op
UnsafeBufferPutIntAsciiBenchmark               27085146  avgt   30  10.538 ± 0.021  ns/op
UnsafeBufferPutIntAsciiBenchmark             -123456789  avgt   30  11.048 ± 0.016  ns/op
UnsafeBufferPutIntAsciiBenchmark             2147483647  avgt   30  11.497 ± 0.057  ns/op
UnsafeBufferPutLongAsciiBenchmark  -9223372036854775808  avgt   30   8.134 ± 0.039  ns/op
UnsafeBufferPutLongAsciiBenchmark                     0  avgt   30   3.275 ± 0.023  ns/op
UnsafeBufferPutLongAsciiBenchmark                 -9182  avgt   30   8.337 ± 0.031  ns/op
UnsafeBufferPutLongAsciiBenchmark              97385146  avgt   30  11.314 ± 0.060  ns/op
UnsafeBufferPutLongAsciiBenchmark     -6180362504315475  avgt   30  19.583 ± 0.207  ns/op
UnsafeBufferPutLongAsciiBenchmark   9223372036854775807  avgt   30  22.412 ± 0.268  ns/op
```
</details>

<details><summary>Lookup table + pre-computed base offset</summary>

```
Benchmark                                       (value)  Mode  Cnt   Score   Error  Units
UnsafeBufferPutIntAsciiBenchmark            -2147483648  avgt   30   6.550 ± 0.021  ns/op
UnsafeBufferPutIntAsciiBenchmark                      0  avgt   30   3.278 ± 0.008  ns/op
UnsafeBufferPutIntAsciiBenchmark                  -9182  avgt   30   7.968 ± 0.039  ns/op
UnsafeBufferPutIntAsciiBenchmark               27085146  avgt   30   9.990 ± 0.040  ns/op
UnsafeBufferPutIntAsciiBenchmark             -123456789  avgt   30  11.646 ± 0.078  ns/op
UnsafeBufferPutIntAsciiBenchmark             2147483647  avgt   30  11.617 ± 0.079  ns/op
UnsafeBufferPutLongAsciiBenchmark  -9223372036854775808  avgt   30   8.065 ± 0.016  ns/op
UnsafeBufferPutLongAsciiBenchmark                     0  avgt   30   3.290 ± 0.010  ns/op
UnsafeBufferPutLongAsciiBenchmark                 -9182  avgt   30   8.597 ± 0.091  ns/op
UnsafeBufferPutLongAsciiBenchmark              97385146  avgt   30  11.165 ± 0.067  ns/op
UnsafeBufferPutLongAsciiBenchmark     -6180362504315475  avgt   30  19.266 ± 0.061  ns/op
UnsafeBufferPutLongAsciiBenchmark   9223372036854775807  avgt   30  22.187 ± 0.074  ns/op
```
</details>